requirements.txt
streamlit
python-dotenv
pillow
pypdf2
nltk
transformers